Update HaloTeamProblem
Modifed team loops after realizing that we skipped potential matches. The "triangle" theory was that if we already tested A=1, B=2, C=3, and D=4, there's no need to test A=2, B=1, C=3, D=4. So if loop A is from 0 to totCombos, B can start at A+1, and C can start at B+1 and D can start at C+1.  But, it would appear Sets are unordered in python, so my assumption is incorrect here.  An optimization could be done to remove the sets in favor of ordered array, but then we lose the ease of testing sets (for more nested fore loops)

So instead of 
A = 0 to TotCombos-1
B = A+1 to TotCombos-1
C = B+1 to TotCombos-1
D = C+1 to TotCombos-1

we have
A = 0 to TotCombos-1
B = 0 to TotCombos-1
C = 0 to TotCombos-1
D = 0 to TotCombos-1