Update HaloTeamProblem
Changes made allow for user to give a minimum score and return all options within that threshold.  This request arose after logic was returning team combinations with the best and worst player on different teams, leading to questions. This was the optimal solution of 1.97, but there was an option with the 2 players on the same team, and they had a 1.98 score.  User would like the option to pick from a list of optimal solutions.

use maxScore to return all team combos with scores less than this number.

use maxOptions to limit options to a specific count. Output can be less than this number if maxScore isn't high enough to give enough options.

Other changes include changes to debug output info